# public/api-docs.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Wavecrest Pro API",
    "version": "1.0.0",
    "description": "Wavecrest Pro is a real-time trend intelligence platform that aggregates trending topics across multiple social platforms. This API provides endpoints for authentication, trend discovery, dashboard management, billing, and analytics.",
    "contact": {
      "name": "Wavecrest Pro Support",
      "url": "https://wavecrest.pro",
      "email": "support@wavecrest.pro"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://wavecrest.pro",
      "description": "Production server"
    }
  ],
...
  "paths": {
    "/api/auth/signup": {
      "post": {
        "tags": ["Auth"],
        "summary": "Create a new account",
        "description": "Register a new user account with email and password. A session cookie is automatically set on success.",
        "operationId": "signup",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequest"
              },
              "example": {
                "email": "you@example.com",
                "password": "secret123",
                "name": "Efe"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                },
                "example": {
                  "message": "Account created successfully",
                  "user": {
                    "id": 1,
                    "email": "you@example.com",
                    "name": "Efe",
                    "plan": "free",
                    "provider": "local",
                    "created_at": "2025-01-15T10:30:00.000Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error (missing fields or email already exists)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Email already registered"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Log in to an existing account",
        "description": "Authenticate with email and password. A session cookie is set on success.",
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              },
              "example": {
                "email": "you@example.com",
                "password": "secret123"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                },
                "example": {
                  "message": "Login successful",
                  "user": {
                    "id": 1,
                    "email": "you@example.com",
                    "name": "Efe",
                    "plan": "pro",
                    "provider": "local",
                    "created_at": "2025-01-15T10:30:00.000Z"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Invalid email or password"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Log out the current session",
        "description": "Destroy the current session and clear the session cookie.",
        "operationId": "logout",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": {
            "description": "Logged out successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" }
                  }
                },
                "example": {
                  "message": "Logged out successfully"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Get current user info",
        "description": "Returns the currently authenticated user's profile and plan information.",
        "operationId": "getMe",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": {
            "description": "Current user info",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                },
                "example": {
                  "id": 1,
                  "email": "you@example.com",
                  "name": "Efe",
                  "plan": "pro",
                  "provider": "local",
                  "created_at": "2025-01-15T10:30:00.000Z"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Not authenticated"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/{provider}/callback": {
      "post": {
        "tags": ["Auth"],
        "summary": "OAuth provider callback",
        "description": "Handle OAuth callback from a third-party provider (e.g., Google, GitHub). Exchanges the authorization code for a session.",
        "operationId": "oauthCallback",
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "description": "OAuth provider name",
            "schema": {
              "type": "string",
              "enum": ["google", "github"]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["code"],
                "properties": {
                  "code": {
                    "type": "string",
                    "description": "Authorization code from OAuth provider"
                  }
                }
              },
              "example": {
                "code": "4/0AX4XfWh..."
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OAuth login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                },
                "example": {
                  "message": "Login successful",
                  "user": {
                    "id": 2,
                    "email": "you@gmail.com",
                    "name": "Efe",
                    "plan": "free",
                    "provider": "google",
                    "created_at": "2025-01-16T08:00:00.000Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired authorization code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Invalid authorization code"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Session cookie set after login or signup"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "description": "Unique user ID" },
          "email": { "type": "string", "format": "email" },
          "name": { "type": "string" },
          "plan": { "type": "string", "enum": ["free", "plus", "pro", "max", "teams", "enterprise"] },
          "provider": { "type": "string", "enum": ["local", "google", "github"] },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "SignupRequest": {
        "type": "object",
        "required": ["email", "password", "name"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 6 },
          "name": { "type": "string" }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "message": { "type": "string" },
          "user": { "$ref": "#/components/schemas/User" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "description": "Human-readable error message" }
        }
      },
...
    }
  }
}